+++ packages/backend-dom/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @glissade/backend-dom
 
+## 0.32.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @glissade/scene@0.32.0-pre.1
+  - @glissade/core@0.32.0-pre.1
+
 ## 0.32.0-pre.0
 
 ### Patch Changes