+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @glissade/cli
 
+## 0.32.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @glissade/scene@0.32.0-pre.1
+  - @glissade/core@0.32.0-pre.1
+  - @glissade/backend-skia@0.32.0-pre.1
+  - @glissade/interact@0.32.0-pre.1
+  - @glissade/lottie@0.32.0-pre.1
+  - @glissade/narrate@0.32.0-pre.1
+  - @glissade/player@0.32.0-pre.1
+  - @glissade/svg@0.32.0-pre.1
+  - @glissade/sfx@0.32.0-pre.1
+
 ## 0.32.0-pre.0
 
 ### Patch Changes