+++ packages/element/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @glissade/element
 
+## 0.32.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @glissade/scene@0.32.0-pre.1
+  - @glissade/core@0.32.0-pre.1
+  - @glissade/backend-canvas2d@0.32.0-pre.1
+  - @glissade/player@0.32.0-pre.1
+
 ## 0.32.0-pre.0
 
 ### Patch Changes