+++ packages/sfx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @glissade/sfx
 
+## 0.32.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/core@0.32.0-pre.1
+
 ## 0.32.0-pre.0
 
 ### Patch Changes